Added snapshot tests for offset-to-node mapping
